+++ Cargo.toml
@@ -138,6 +138,7 @@ rspack_loader_react_refresh            = { version = "0.2.0", path = "crates/rsp
 rspack_loader_runner                   = { version = "0.2.0", path = "crates/rspack_loader_runner" }
 rspack_loader_swc                      = { version = "0.2.0", path = "crates/rspack_loader_swc" }
 rspack_loader_testing                  = { version = "0.2.0", path = "crates/rspack_loader_testing" }
+rspack_location                        = { version = "0.2.0", path = "crates/rspack_location" }
 rspack_macros                          = { version = "0.2.0", path = "crates/rspack_macros" }
 rspack_napi                            = { version = "0.2.0", path = "crates/rspack_napi" }
 rspack_napi_macros                     = { version = "0.2.0", path = "crates/rspack_napi_macros" }
