+++ crates/rspack_core/Cargo.toml
@@ -42,6 +42,7 @@ rspack_hash = { workspace = true }
 rspack_hook = { workspace = true }
 rspack_javascript_compiler = { workspace = true }
 rspack_loader_runner = { workspace = true }
+rspack_location = { workspace = true }
 rspack_macros = { workspace = true }
 rspack_napi = { workspace = true, optional = true }
 rspack_paths = { workspace = true }
